Исправил замечания.

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -1,6 +1,7 @@
 package ru.hogwarts.school.service;
 
 import org.springframework.stereotype.Service;
+import ru.hogwarts.school.exception.EntityNotFoundException;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
 import ru.hogwarts.school.repository.FacultyRepository;
@@ -19,19 +20,33 @@ public class FacultyService {
     }
 
     public Faculty addFaculties(Faculty faculty) {
+
         return facultyRepository.save(faculty);
     }
 
     public Faculty findFaculty(Long id) {
-        return facultyRepository.getReferenceById(id);
+        Faculty facultyFind = facultyRepository.getReferenceById(id);
+        if (facultyFind == null) {
+            throw new EntityNotFoundException();
+        }
+        return facultyFind;
     }
 
     public Faculty findByName(String name) {
+        Faculty nameByFind = facultyRepository.findByNameIgnoreCase(name);
+        if (nameByFind == null) {
+            throw new EntityNotFoundException();
+        }
         return facultyRepository.findByNameIgnoreCase(name);
     }
 
     public Collection<Student> getAllStudentsFaculty(Long id) {
-        return facultyRepository.getReferenceById(id).allStudentsFaculty();
+        Collection<Student> allStudent =
+                facultyRepository.getReferenceById(id).allStudentsFaculty();
+        if (allStudent.isEmpty()) {
+            throw new EntityNotFoundException();
+        }
+        return allStudent;
     }
 
     public Collection<Faculty> getAll() {
@@ -44,13 +59,16 @@ public class FacultyService {
 
     public Faculty editFaculty(Faculty faculty) {
         if (facultyRepository.getReferenceById(faculty.getId()) == null) {
-            return null;
+            throw new EntityNotFoundException();
         }
         return facultyRepository.save(faculty);
     }
 
     public Faculty removeFaculty(Long id) {
         Faculty facultyRemove = facultyRepository.getReferenceById(id);
+        if (facultyRemove == null) {
+            throw new EntityNotFoundException();
+        }
         facultyRepository.delete(facultyRemove);
         return facultyRemove;
     }
